fix/update: fix circular dependency in useTokenLists. Seperate some of the concerns and update usages (#465)
