+++ src/components/cards/TradeCard/TradeRoute.vue
@@ -129,7 +129,7 @@ import useNumbers from '@/composables/useNumbers';
 import { SorReturn } from '@/lib/utils/balancer/helpers/sor/sorManager';
 import { useI18n } from 'vue-i18n';
 import useVueWeb3 from '@/services/web3/useVueWeb3';
-import useTokenLists from '@/composables/useTokenLists';
+import useTokens from '@/composables/useTokens';
 
 interface Route {
   share: number;
@@ -183,7 +183,7 @@ export default defineComponent({
     const { t } = useI18n();
 
     const { userNetworkConfig } = useVueWeb3();
-    const { tokenDictionary: tokens } = useTokenLists();
+    const { tokens } = useTokens();
 
     const visible = ref(false);
