+++ src/components/modals/TradePreviewModal.vue
@@ -84,8 +84,8 @@ import { defineComponent, toRefs, computed } from 'vue';
 import { ETHER } from '@/constants/tokenlists';
 import useNumbers from '@/composables/useNumbers';
 import useTokenApproval from '@/composables/trade/useTokenApproval';
-import useTokenLists from '@/composables/useTokenLists';
 import useVueWeb3 from '@/services/web3/useVueWeb3';
+import useTokens from '@/composables/useTokens';
 
 export default defineComponent({
   emits: ['trade', 'close'],
@@ -124,7 +124,7 @@ export default defineComponent({
 
     const { addressIn, amountIn, addressOut, isV1Swap } = toRefs(props);
 
-    const { tokenDictionary: tokens } = useTokenLists();
+    const { tokens } = useTokens();
     const { userNetworkConfig } = useVueWeb3();
 
     const isWrap = computed(() => {
